Flytter ut metoden for å oppdatere grunnlaget til egen komponent for ryddigere komponent
